+++ NEWS.md
@@ -1,4 +1,4 @@
-# adaR (development version)
+# adaR 0.4.0
 
 ## Breaking changes